+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- setup: Don't install `micromamba` on the raspi OS image
 - tools: Remove legacy Wi-Fi access point setup script
 - tools: Remove update command from `upkie_tool`